+++ src/rbm.h
@@ -5,8 +5,8 @@
 
 typedef struct {
   // Basic description of a restricted Boltzman machine.
-  long n_inputs;           // Number of nodes in the lower 'input' layer.
-  long n_outputs;          // Number of nodes in the upper 'output' layer.
+  long long n_inputs;           // Number of nodes in the lower 'input' layer.
+  long long n_outputs;          // Number of nodes in the upper 'output' layer.
   
   matrix_t *io_weights;   // n_inputs x n_outputs matrix representing weights between input and output nodes.
   double *bias_inputs;    // 1x n_inputs vector representing the 'bias' toward activation of the input nodes.
